feat: M6a Zp 类 64-bit 扩展（FLINT 归一化 Barrett）
- Zp 数据成员 uint32 → uint64，新增 _norm 字段（对象 24B → 32B）
- 乘法：FLINT 归一化 Barrett 归约（__nmod_mul，~15 cycles）
- 预计算：__preinvert_limb + __precompute 替换 __barrett_ninv
- inv_prime：__int128 防溢出 + 条件减法
- 下游适配：10 个源文件 + 10 个测试文件 uint32→uint64
- mtshl_p 升级为最大 63-bit 素数 9223372036854775783
- 新增 4 个 63-bit 素数测试（算术/边界/乘法压力/除法+pow）

+++ clpoly/polynomial.hh
@@ -674,7 +674,7 @@ namespace clpoly{
     }
 
     template <class compare>
-    polynomial_<Zp,compare> polynomial_mod(const polynomial_<ZZ,compare> & p, uint32_t prime)
+    polynomial_<Zp,compare> polynomial_mod(const polynomial_<ZZ,compare> & p, uint64_t prime)
     {
         polynomial_<Zp,compare> new_p(p.comp_ptr());
         // std::cout<<p<<" mod "<<prime;
@@ -689,7 +689,7 @@ namespace clpoly{
         return new_p;
     }
     template <class compare>
-    polynomial_<Zp,compare> polynomial_mod(polynomial_<ZZ,compare> && p, uint32_t prime)
+    polynomial_<Zp,compare> polynomial_mod(polynomial_<ZZ,compare> && p, uint64_t prime)
     {
         polynomial_<Zp,compare> new_p(p.comp_ptr());
         Zp coeff(prime);
